requirements.txt
transformers
pillow
torch 
torchvision
fastapi
python-multipart
'uvicorn[standard]'